2014 04 27   17:46

+++ BD2_2M/T3 - Avalicação 2014-1/Produtos.sql
@@ -4,7 +4,7 @@ CREATE TABLE produtos (
 	nome VARCHAR(50) UNIQUE NOT NULL,
 	unidade INTEGER, 
 	quantidade INTEGER NOT NULL, 
-	preco_unitario DOUBLE, 
+	preco_unitario DOUBLE PRECISION, 
 	estoque_minimo INTEGER,
 	estoque_maximo INTEGER UNIQUE
